\name{citations}
\alias{citations}
\title{Citations With All Entries Shown}
\usage{
citations(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
The input, invisibly.
}
\description{
Follows \bibcitet{R:Smith2020}. See \href{#details}{details},
\url{https://cran.r-project.org/package=edgepkg},
\href{mailto:jane.doe@gmail.com}{the maintainer} and \href{www.r-project.org/about.html}{R}.
}
\references{
\bibshow{*}
}
